requirements.txt
appdirs==1.4.4
astor==0.8.1
bracex==2.2.1
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
csscompressor==0.9.5
cssselect2==0.6.0
defusedxml==0.7.1
funcparserlib==1.0.0
ghp-import==2.1.0
gitdb==4.0.9
GitPython==3.1.27
htmlmin==0.1.12
hy==0.20.0
idna==3.3
importlib-metadata==4.11.3
Jinja2==3.0.3
jsmin==3.0.1
lxml==4.8.0
macros==0.0.1
Markdown==3.3.6
MarkupSafe==2.1.1
mergedeep==1.3.4
mkdocs==1.3.0
mkdocs-awesome-pages-plugin==2.7.0
mkdocs-badges==0.3.2
mkdocs-macros-plugin==0.7.0
mkdocs-material-extensions==1.0.3
mkdocs-minify-plugin==0.5.0
mkdocs-redirects==1.0.4
mkdocs-snippet-plugin==1.0.2
numpy==1.23.0rc2
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pycparser==2.21
Pygments==2.12.0
pymdown-extensions==9.4
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyyaml_env_tag==0.1
regex==2022.4.24
requests==2.27.1
rply==0.7.8
six==1.16.0
smmap==5.0.0
termcolor==1.1.0
tinycss2==1.1.1
urllib3==1.26.9
watchdog==2.1.7
wcmatch==8.3
webencodings==0.5.1
zipp==3.8.0
